wip arch

+++ install.sh
@@ -1,7 +1,16 @@
 #!/usr/bin/env bash
-# execute with bash -c "$(wget -O - https://raw.githubusercontent.com/grota/chezmoitest/master/install.sh)"
+# execute with bash -c "$(wget -O - https://raw.githubusercontent.com/grota/dotfiles/master/install.sh)"
 
 set -e
+
+exists() {
+  if command -v "$1" >/dev/null 2>&1
+  then
+    return 0
+  else
+    return 1
+  fi
+}
 
 BIN_DIR="$HOME/local/bin"
 
@@ -11,10 +20,10 @@ fi
 
 export PATH="$PATH:$BIN_DIR"
 
-if [ ! "$(command -v chezmoi)" ]; then
-  if [ "$(command -v curl)" ]; then
+if ! exists chezmoi; then
+  if exists curl; then
     sh -c "$(curl -fsSL https://git.io/chezmoi)" -- -b "$BIN_DIR"
-  elif [ "$(command -v wget)" ]; then
+  elif exists wget; then
     sh -c "$(wget -qO- https://git.io/chezmoi)" -- -b "$BIN_DIR"
   else
     echo "To install chezmoi, you must have curl or wget installed." >&2
@@ -25,7 +34,11 @@ fi
 
 if [ ! "$(command -v bw)" ]; then
   TMPFILE=$(mktemp)
-  wget -q -O "$TMPFILE" "https://vault.bitwarden.com/download/?app=cli&platform=linux"
+  if exists wget; then
+    wget -q -O "$TMPFILE" "https://vault.bitwarden.com/download/?app=cli&platform=linux"
+  else
+    curl -fsSL -o "$TMPFILE" "https://vault.bitwarden.com/download/?app=cli&platform=linux"
+  fi
   unzip "$TMPFILE" -d "$BIN_DIR"
   rm "$TMPFILE"
   chmod u+x "$BIN_DIR/bw"
@@ -36,3 +49,4 @@ echo "Login with bitwarden cli."
 read -p "Your Bitwarden email: " BW_EMAIL
 read -p "Your Bitwarden master password: " BW_PASSWORD
 bw login "$BW_EMAIL" "$BW_PASSWORD"
+echo "These dotfiles assume ~/.bash_aliases is sourced by ~/.bashrc, fix it if it is not so."
